django-backend/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-elasticsearch-dsl-drf==0.20.8
django-cors-headers==3.2.1
requests==2.23.0